First attempt at collision
In this one I decided to see if the player was in a platform tile or not.
However, there are some strange bugs when it comes to left + right.
Right now, if the player is touching a tile on the ground it counts as left + right as well.

We do need to get the player's position on the map.
However, we also need to check the rectangles of the platforms with the player.