Add a level class
We're not adding classes for the sake of them.
In every level we have the map used to display tiles.
We also have the dimensions that should be drawn on the screen.
We want to store the start position so we can respawn the player.

Since we need these properties for many levels, make a class for it.